refactor(ui): extract tasks/useBackgroundTasks hook from App.tsx
Moves the background-task drawer state (open/selected/list/output/clock),
the list+output polling loops, and the by-tool-call index into a dedicated
hook keyed by sessionId. Pure move, no behavior change.
